Lambda's source variable is now either a single GeneralVariable, or a Tuple.  Could be improved so that only X's are acceptable where X's are either GeneralVariables or Tuples of X's.
Also, Reworked documentation for Quotient.

Also, Added "Unspecified" as a Map constructor.  It's domain is the UnitSpace, i.e. it is essentially a Value.  The intention is that when Connectivity is specified (i.e. by forming a Quotient Topological space), the target to which some (often most) points map is often left "Unspecified", meaning that they are not connected. This should've been split into a separate commit, lumped in due to laziness.