WIP continued, and still WIP: fixing misconception of Lambdas and interplay with non-Lambda expressions.
Renamed Map to Expression, and related renames.
Some more renaiming from space to FSet.

Also, added in "Lambdify", which will allow the original concept of treating naked expressions as Lambdas to be restored.  The idea is that
codomain expr1 will only make sense if expr1 is a lambda, but the previous usage will be available via codomain (Lambdify expr1).
This is still very much considered experimental.