WIP continued, and still WIP: fixing misconception of Lambdas and interplay with non-Lambda expressions.
Mainly addressed listOfFreeGeneralVariables to be consistent with correct view of expressions vs Lambdas.

Also, some minor updates to documentation, trying to vaguely stick to a convention of using
x when the expression is expected to be a variable or variable tuple,
expr1 if it is expected to be an expression, and
f or g if it is expected to be a Lambda.